+++ src/binary/boxed.rs
@@ -1,19 +1,21 @@
 // Copyright (C) 2025 FZI Forschungszentrum Informatik
 // SPDX-License-Identifier: Apache-2.0
-//! [`Binary`] requiring [`Box`] and other types from alloc [`alloc`]
+//! [`Binary`][BinTrait] requiring [`Box`] and other types from alloc [`alloc`]
 
 use alloc::boxed::Box;
+use core::fmt;
 
 use crate::instruction::{Instruction, info};
 
-use super::Binary;
-use super::error::MaybeMissError;
+use super::Binary as BinTrait;
+use super::error;
 
-/// [`Binary`] returning a boxed, dynamically dispatched `Error`
+/// [`Binary`][BinTrait] returning a boxed, dynamically dispatched [`Error`]
 ///
-/// This [`Binary`] adapter boxes and type-erases errors returned by the wrapped
-/// [`Binary`]. This allows dynamically dispatching [`Binary`]s with differrent
-/// [`Binary::Error`] types.
+/// This [`Binary`][BinTrait] adapter boxes and type-erases errors returned by
+/// the wrapped [`Binary`][BinTrait]. This allows dynamically dispatching
+/// [`Binary`][BinTrait]s with differrent [`Binary::Error`][BinTrait::Error]
+/// types.
 #[derive(Copy, Clone, Debug)]
 pub struct BoxedError<B> {
     inner: B,
@@ -32,17 +34,52 @@ impl<B> From<B> for BoxedError<B> {
     }
 }
 
-impl<B, I> Binary<I> for BoxedError<B>
+impl<B, I> BinTrait<I> for BoxedError<B>
 where
-    B: Binary<I>,
-    B::Error: MaybeMissError + 'static,
+    B: BinTrait<I>,
+    B::Error: error::MaybeMissError + 'static,
     I: info::Info,
 {
-    type Error = Box<dyn MaybeMissError>;
+    type Error = Error;
 
     fn get_insn(&mut self, address: u64) -> Result<Instruction<I>, Self::Error> {
-        self.inner
-            .get_insn(address)
-            .map_err(|e| -> Box<dyn MaybeMissError> { Box::new(e) })
+        self.inner.get_insn(address).map_err(|e| Box::new(e).into())
     }
 }
+
+/// Dynamically dispatched error
+#[derive(Debug)]
+pub struct Error(Box<dyn error::MaybeMissError>);
+
+impl<E: error::MaybeMissError + 'static> From<Box<E>> for Error {
+    fn from(err: Box<E>) -> Self {
+        Self(err)
+    }
+}
+
+impl error::Miss for Error {
+    fn miss(address: u64) -> Self {
+        Self(Box::new(error::NoInstruction::miss(address)))
+    }
+}
+
+impl error::MaybeMiss for Error {
+    fn is_miss(&self) -> bool {
+        error::MaybeMiss::is_miss(self.0.as_ref())
+    }
+}
+
+impl core::error::Error for Error {
+    fn source(&self) -> Option<&(dyn core::error::Error + 'static)> {
+        core::error::Error::source(self.0.as_ref())
+    }
+}
+
+impl fmt::Display for Error {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        fmt::Display::fmt(self.0.as_ref(), f)
+    }
+}
+
+/// a [`Binary`][BinTrait] boxed for dynamic dispatch
+pub type Binary<'a, I> = Box<dyn BinTrait<I, Error = Error> + 'a>;